Replace original csv python script with new script which returns all fields.
